Revert "[Proposal-1] Support Partitioner Refactoring (#845)"
This reverts commit a8d9f1d25fbb764b12ce41ba28bc15eb25786859.
