+++ src/main/java/org/apache/pulsar/io/jcloud/BlobStoreAbstractConfig.java
@@ -42,7 +42,6 @@ import org.apache.pulsar.io.jcloud.format.Format;
 import org.apache.pulsar.io.jcloud.format.JsonFormat;
 import org.apache.pulsar.io.jcloud.format.ParquetFormat;
 import org.apache.pulsar.io.jcloud.partitioner.PartitionerType;
-import org.apache.pulsar.io.jcloud.partitioner.legacy.LegacyPartitionerType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -82,9 +81,7 @@ public class BlobStoreAbstractConfig implements Serializable {
 
     private String formatType;
 
-    @Deprecated // Use partitioner instead
     private String partitionerType;
-    private PartitionerType partitioner = PartitionerType.LEGACY;
 
     private boolean partitionerUseIndexAsOffset;
 
@@ -136,18 +133,18 @@ public class BlobStoreAbstractConfig implements Serializable {
         }
 
         if (partitionerType == null
-                || (EnumUtils.getEnumIgnoreCase(LegacyPartitionerType.class, partitionerType) == null
+                || (EnumUtils.getEnumIgnoreCase(PartitionerType.class, partitionerType) == null
                 && !partitionerType.equalsIgnoreCase("default"))) {
             // `default` option is for backward compatibility
             throw new IllegalArgumentException(
                     "partitionerType property not set properly, available options: "
-                            + Arrays.stream(LegacyPartitionerType.values())
+                            + Arrays.stream(PartitionerType.values())
                             .map(Enum::name)
                             .map(String::toLowerCase)
                             .collect(Collectors.joining(","))
             );
         }
-        if (LegacyPartitionerType.TIME.name().equalsIgnoreCase(partitionerType)) {
+        if (PartitionerType.TIME.name().equalsIgnoreCase(partitionerType)) {
             if (StringUtils.isNoneBlank(timePartitionPattern)) {
                 LOGGER.info("test timePartitionPattern is ok {} {}",
                         timePartitionPattern,
