+++ src/main/java/org/apache/pulsar/io/jcloud/partitioner/AbstractPartitioner.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.io.jcloud.partitioner.legacy;
+package org.apache.pulsar.io.jcloud.partitioner;
 
 import java.util.ArrayList;
 import java.util.List;
