+++ src/test/java/org/apache/pulsar/io/jcloud/partitioner/SliceTopicPartitionPartitionerTest.java
@@ -29,9 +29,6 @@ import org.apache.pulsar.client.impl.MessageIdImpl;
 import org.apache.pulsar.common.naming.TopicName;
 import org.apache.pulsar.functions.api.Record;
 import org.apache.pulsar.io.jcloud.BlobStoreAbstractConfig;
-import org.apache.pulsar.io.jcloud.partitioner.legacy.Partitioner;
-import org.apache.pulsar.io.jcloud.partitioner.legacy.SimplePartitioner;
-import org.apache.pulsar.io.jcloud.partitioner.legacy.TimePartitioner;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
